`getRepoStatus` appears to be a performance bottleneck, as it's called on almost all `git-meta` commands.  This PR addresses some of the performance problem by removing assertions from critical paths.

On my laptop, with a meta-repo containing 6,000 repositories, this dropped the time taken to call `getRepoStatus` from ~1,500ms to ~450ms and overall wall time from ~1.9s to ~0.8s.